Implementation of AddTenant Class
The addTenant dialog prompt has been replaced with a class to allow for action listeners to be added. This ensures that if both name fields aren't filled the tenant won't be created. tenantList now updates with added/edited tenants  when Save button is pressed. Still need to get tenantList displaying email and phone but I have an idea of how to go about this.