Delete button
Moved delete button to primary GUI panel as cross class functions were causing major issues. The method simply checks the index of tenantList and if there's properties, it will remove the current selected tenant from the list (just like removeProperty). EOFexception is occuring when attempting to delete the tenant from the dB. I was unable to figure out what is causing this.